backend/pyproject.toml
[project]
name = "gemini-motion-lab-backend"
version = "0.1.0"
description = "Backend for Gemini Motion Lab"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.2.0",
    "google-genai>=1.0.0",
    "google-cloud-storage>=2.16.0",
    "opencv-python-headless>=4.9.0",
    "imageio-ffmpeg>=0.4.9",
    "mediapipe>=0.10.0",
    "python-dotenv>=1.0.0",
    "cryptography>=42.0.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []
